Duration pointer is now 32 bit.
Had to do some math to pad out the various loops to prevent vibrato. I
think absolute pitch is slightly off, but there is no vibrato effect
that fades in and out (something that was occurring when I didn't pad
the assembly that checks the duration values with nop calls).